Added initial code for testing USB design
The USB design is a simple ATTiny85 connected to a USB port, along with
a reset button.  Power regulation is provided via two Zener diodes, and
the Micronucleous USB bootloader is flashed.

This simple example program just flashes an LED on PRB0 of the ATtiny85,
and is installed via micronucleus.