MusicalBook: Successful test complete
This completes a successful test of the initial MusicalBook board.
Although the test was successful, a fault was identified (ATMega GND not
connected to GND).  Aside from that, it appears to work well for this
test.

The test itself simply uploads an LED blinker, with the LED connected to
the SPEAKER output terminal.  The test doesn't check any of the input
terminals, so that remains to be done.